requirements.txt
flask
pandas
# pymongo
Flask-Redis
click
